pyproject.toml
[project]
name = "geojson-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.0",
    "geopandas>=1.1.1",
    "google-genai>=1.25.0",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "streamlit>=1.46.1",
    "uvicorn>=0.35.0",
    "pandas>=2.2.2",
    "openpyxl>=3.1.5",
    "fastkml>=0.4",
    "lxml>=6.0.0",
]
